+++ lib/src/icons/between_horizontal_end_icon.dart
@@ -14,6 +14,9 @@ class BetweenHorizontalEndIcon extends AnimatedSVGIcon {
     super.enableTouchInteraction = true,
     super.infiniteLoop = false,
     super.resetToStartOnComplete = true,
+    super.onTap,
+    super.interactive,
+    super.controller,
   });
 
   @override