+++ lib/src/icons/bomb_icon.dart
@@ -15,6 +15,9 @@ class BombIcon extends AnimatedSVGIcon {
     super.enableTouchInteraction = true,
     super.infiniteLoop = false,
     super.resetToStartOnComplete = true,
+    super.onTap,
+    super.interactive,
+    super.controller,
   });
 
   @override